Ensure clean shutdown of monitor
Make default SIGTERM and SIGINT handlers use tevent

Wait for all children to exit

Previously, there was a race-condition where the monitor might
terminate before its children.

Resend SIGTERM if child doesn't terminate

There is a race condition where if we send a SIGTERM before the
kernel has scheduled the child, it may be lost, and the child will
not terminate and will leave the monitor hung in wait().

This patch alters this behavior so that we will send the SIGTERM
again every 10ms and check the wait() in a nonblocking manner.

Set up signal handlers before initializing sysdb

A temporary signal handler for SIGTERM is set up in server_setup()
that calls exit() from within a pure signal handler. This causes a
race condition where it's possible that if the SSSD is restarted
immediately while it is still initializing the sysdb caches for
the first time, it can leave the cache in a corrupt and unusable
state.

https://bugzilla.redhat.com/show_bug.cgi?id=658444