Do not attempt to resolve nameless servers
The failover code is not strictly in charge of resolving. Its main
function is to provide a server to connect to for a service.

It is legal, although not currently used, to have a server that has no
name (server->common == NULL). In this case, no resolving should be done
and it is assumed that the failover user, which are the SSSD back ends
in our case, would perform any resolving out of band, perhaps using the
user_data attribute of fo_server structure.