Return pam data to the renewal item if renewal fails
A previous patch changed a talloc_steal() into a talloc_move(). Now it
is not enough to change the parent memory context with talloc_steal to
give back the data, but it has to be assigned back too.

Additionally this patch uses the missing pam data as an indication that
a renewal request for this data is currently running.