Just use the service name with krb5_get_init_creds_password()
Currently we add the realm name to change password principal but
according to the MIT Kerberos docs and the upstream usage the realm name
is just ignored.

Dropping the realm name also does not lead to confusion if the change
password request was received for a user of a trusted domain.