AUTOFS: remove all maps from hash if request for auto.master comes in
https://fedorahosted.org/sssd/ticket/1592

When a request for auto.master comes in, we need to remove all the maps
from the lookup hash table. We can't simply delete the maps, because
another request might be processing them, so instead the maps are
removed from the hash table, effectively becoming orphaned. The maps
will get freed when the timed destructor is invoked.