LDAP SUDO: sudo provider doesn't fetch 'EntryUSN'
The EntryUSN is not fetched by the sudo LDAP provider when it downloads
the rules because sudorule_map is missing this attribute. We forgot to
add the SDAP_AT_SUDO_RUNAS into sdap_sudorule_attrs when we added support
for sudoRunAs.

Related to:
    https://fedorahosted.org/sssd/ticket/2212

(cherry picked from commit a1e89ede4995f948abc1acc364246161df7cca2c)