UTIL: Do not change SSSD domains in get_domains_head
When there was more than one SSSD domain configured, actions performed
against domains later in the list would be incorrectly told to use the
first domain as the base for locating subdomains. This was because we
were rewinding the ->prev list on the sss_domain_info object, which is
only intended to be used by confdb code. The correct approach was to
use only the parent linkage, which would take us up to the top-level
domain in this SSSD domain.

(cherry picked from commit a2147c6c13c36c0ec056581b08a7ed7352ff9861)