nss: preserve service name in getsrv call
About case_sensitive=preserving and services.

The name of the service can be preserved in
result of 'getent service'. However we
should still lowercase the protocol and
service aliases because they serve as keys
in some queries to sysdb. The lowercasing is done
by the provider already. If we did not do that,
we would lose case insesnsitivity.

With this patch the responder preserves the
case of service name and protocol, to match
the case that is stored in the sysdb (however
the protocol is already lowercased by provider,
so it was done only for consistent use of the
case_sensitive=preserve option in the responders
and only the case of name is the same as in
ldap).

Fixes:
https://fedorahosted.org/sssd/ticket/2460

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit c4e278c2637547640d28bda007c9c38a17956fcc)