views: allow view name change at startup
Currently some manual steps are needed on a FreeIPA to switch from one
view to another. With this patch the IPA provider checks at startup if
the view name changed and does the needed steps automatically. Besides
saving the new view name this includes removing the old view data and
marking the user and group entries as invalid.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>