FO: Add an API to reset all servers in a single service
Required for:
    https://fedorahosted.org/sssd/ticket/2639

Previously, we had a function that allowed the caller to reset the
status of all services in the global fail over context. This patch adds
a new function that allows the caller to reset a single service instead.

The main user would be IPA subdomain provider that might need to reset
the status of an AD trusted domain on demand.