FO: Also reset the server common data in addition to SRV
In a server that is expanded from a SRV query was reset, only it's
'meta-server' status was set to neutral, but the server->common
structure still retained its not_working status.

This patch also resets the status of the common structure so that both
the SRV query and resolving the server are retried next time.

Reviewed-by: Sumit Bose <sbose@redhat.com>
(cherry picked from commit 6ac6e8f83da241458742b7f8a3406ed66bded292)