Responders: Fix client destrcutor
To close a socket associated to an fd event we must set the close
function of the event and not associate a destructor to aparent context.

Otherwise the destructor will close() the socket before the fd event is
freed, and this may cause invalid calls on a closed file descriptor to
poll/epoll/etc.

Discovered by looking at startce output.

Resolves:
https://fedorahosted.org/sssd/ticket/2973