AD: replace ad_get_client_site_parse_ndr() with netlogon_get_domain_info()
netlogon_get_domain_info() does not fail if only the site is missing in
the CLDAP ping respond. If the site is not available a Global Catalog
can still be looked up with the forest name. Only if the forest name is
missing as well we fall back to the configured domain name.

Resolves:
https://fedorahosted.org/sssd/ticket/3104

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>
(cherry picked from commit 3000d86bea2bae735175f93b0e013bbcb3979f1e)