KCM: request parsing and sending a reply
Implements parsing the KCM client request into per-client buffers and
sending a response for both the failure case and for success.

The protocol is documented at:
    http://k5wiki.kerberos.org/wiki/Projects/KCM_client

Several places don't use the sss_iobuf structure, because they don't
parse variable-length data from the buffer and it's much more efficient
to just allocate the needed request and reply structure on the stack.