RESPONDER: Add a responder sbus interface to disable domains
Adds a generic responder s-bus interface that all responders implement.
The interface currently contains methods that make it possible for a sssd
domain to be disabled and enabled by a back end.

Disabled domains are not supposed to be even considered by responders
and are just skipped.

In the future, this commit will be superseded by sbus signals.